+++ travis/config/karma/test.karma.conf.js
@@ -13,7 +13,7 @@ function getConfig(config) {
     BROWSER_STACK_USERNAME: process.env.BROWSER_STACK_USERNAME,
     BROWSER_STACK_ACCESS_KEY: process.env.BROWSER_STACK_ACCESS_KEY,
     BROWSER_STACK_BUILD_ID: process.env.TRAVIS_BUILD_NUMBER,
-    BROWSER_STACK_PROJECT: 'SKY UX Travis CI'
+    BROWSER_STACK_PROJECT: 'SKY UX Travis CI: Karma'
   });
 }
